requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asttokens==3.0.1
avro==1.11.3
backoff==2.2.1
blinker==1.9.0
boto3==1.42.43
botocore==1.42.43
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.4
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.5
dagshub-annotation-converter==0.1.16
databricks-sdk==0.88.0
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
dnspython==2.8.0
docker==7.1.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.129.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2025.10.0
furl==2.1.4
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
hopsworks==4.2.10
hopsworks-aiomysql==0.2.1
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==8.38.0
itsdangerous==2.2.0
javaobj-py3==0.4.4
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mock==5.2.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opensearch-py==2.4.2
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orderedmultidict==1.0.2
packaging==25.0
pandas==2.1.4
parso==0.8.6
pathvalidate==3.3.1
pillow==12.1.1
platformdirs==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyhumps==1.6.1
pyjks==20.0.0
pymongo==4.16.0
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.3.2
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.29
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
tenacity==9.1.4
threadpoolctl==3.6.0
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
treelib==1.8.0
twofish==0.3.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.5
yarl==1.22.0
zipp==3.23.0
